+++ src/Plutarch/Extra/Maybe.hs
@@ -12,6 +12,9 @@ module Plutarch.Extra.Maybe (
     -- * Utility functions for working with 'PMaybeData'
     pfromDJust,
     pisDJust,
+    pmaybeData,
+    pdjust,
+    pdnothing,
 
     -- * Conversion between 'PMaybe' and 'PMaybeData'
     pmaybeToMaybeData,
@@ -25,7 +28,13 @@ module Plutarch.Extra.Maybe (
 ) where
 
 import Data.Kind (Type)
-import Plutarch (
+import Plutarch.Api.V1.Maybe (PMaybeData (PDJust, PDNothing))
+import Plutarch.Bool (PBool)
+import Plutarch.Builtin (PIsData, pdata, pfromData)
+import Plutarch.DataRepr (pdcons, pdnil, pfield)
+import Plutarch.Lift (pconstant)
+import Plutarch.Maybe (PMaybe (PJust, PNothing))
+import Plutarch.Prelude (
     PCon (pcon),
     S,
     Term,
@@ -33,14 +42,9 @@ import Plutarch (
     plam,
     pmatch,
     (#),
+    (#$),
     type (:-->),
  )
-import Plutarch.Api.V1.Maybe (PMaybeData (PDJust, PDNothing))
-import Plutarch.Bool (PBool)
-import Plutarch.Builtin (PIsData, pdata, pfromData)
-import Plutarch.DataRepr (pdcons, pdnil, pfield)
-import Plutarch.Lift (pconstant)
-import Plutarch.Maybe (PMaybe (PJust, PNothing))
 import Plutarch.String (PString)
 import Plutarch.TermCont (TermCont, tcont)
 import Plutarch.Trace (ptraceError)
@@ -126,6 +130,40 @@ pisDJust = phoistAcyclic $
     plam $ \x -> pmatch x $ \case
         PDJust _ -> pconstant True
         _ -> pconstant False
+
+{- | Special version of 'pmaybe' that works with 'PMaybeData'
+
+ @since 1.3.0
+-}
+pmaybeData ::
+    forall (a :: S -> Type) (b :: S -> Type) (s :: S).
+    PIsData a =>
+    Term s (b :--> (a :--> b) :--> PMaybeData a :--> b)
+pmaybeData = phoistAcyclic $
+    plam $ \d f m -> pmatch m $
+        \case
+            PDJust x -> f #$ pfield @"_0" # x
+            _ -> d
+
+{- | Construct a 'PDJust' value
+
+ @since 1.3.0
+-}
+pdjust ::
+    forall (a :: S -> Type) (s :: S).
+    PIsData a =>
+    Term s (a :--> PMaybeData a)
+pdjust = phoistAcyclic $
+    plam $ \x -> pcon $ PDJust $ pdcons @"_0" # pdata x #$ pdnil
+
+{- | Construct a 'PDNothing' value
+
+ @since 1.3.0
+-}
+pdnothing ::
+    forall (a :: S -> Type) (s :: S).
+    Term s (PMaybeData a)
+pdnothing = phoistAcyclic $ pcon $ PDNothing pdnil
 
 --------------------------------------------------------------------------------
 -- Conversion between 'PMaybe' and 'PMaybeData'.