+++ CHANGELOG.md
@@ -14,8 +14,11 @@ This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
 - `pfindOutputsToAddress`, finds all TxOuts sent to an Address.
 - `pfindTxOutDatum`, finds the data corresponding to a TxOut, if there is one.
 - `phasOnlyOneTokenOfCurrencySymbol`, checks if entire value only contain one token of given currency symbol.
-- `pon`, mirroring `Data.Function.on`
-- `pbuiltinUncurry`, mirroring `uncurry`
+- `pon`, mirroring `Data.Function.on`.
+- `pbuiltinUncurry`, mirroring `uncurry`.
+- `pmaybeData`, mirroring `maybe` for `PMaybeData`.
+- `pdjust` for easier construction of `PDJust` value.
+- `pdnothing` for easier construction `PDNothing` value.
 
 ### Modified